refactor: move tmp dir test util to core-test-kit
